+++ StabilityMatrix.Avalonia/App.axaml.cs
@@ -330,7 +330,8 @@ public sealed class App : Application
                         provider.GetRequiredService<InferenceViewModel>(),
                         provider.GetRequiredService<CheckpointsPageViewModel>(),
                         provider.GetRequiredService<CheckpointBrowserViewModel>(),
-                        provider.GetRequiredService<OutputsPageViewModel>()
+                        provider.GetRequiredService<OutputsPageViewModel>(),
+                        provider.GetRequiredService<WorkflowsPageViewModel>()
                     },
                     FooterPages = { provider.GetRequiredService<SettingsViewModel>() }
                 }
@@ -553,7 +554,7 @@ public sealed class App : Application
             .ConfigureHttpClient(c =>
             {
                 c.BaseAddress = new Uri("https://civitai.com");
-                c.Timeout = TimeSpan.FromSeconds(15);
+                c.Timeout = TimeSpan.FromSeconds(30);
             })
             .AddPolicyHandler(retryPolicy);
 
@@ -562,7 +563,7 @@ public sealed class App : Application
             .ConfigureHttpClient(c =>
             {
                 c.BaseAddress = new Uri("https://civitai.com");
-                c.Timeout = TimeSpan.FromSeconds(15);
+                c.Timeout = TimeSpan.FromSeconds(30);
             })
             .AddPolicyHandler(retryPolicy);
 
@@ -579,6 +580,15 @@ public sealed class App : Application
                 serviceProvider =>
                     new TokenAuthHeaderHandler(serviceProvider.GetRequiredService<LykosAuthTokenProvider>())
             );
+
+        services
+            .AddRefitClient<IOpenArtApi>(defaultRefitSettings)
+            .ConfigureHttpClient(c =>
+            {
+                c.BaseAddress = new Uri("https://openart.ai/api/public/workflows");
+                c.Timeout = TimeSpan.FromSeconds(30);
+            })
+            .AddPolicyHandler(retryPolicy);
 
         // Add Refit client managers
         services.AddHttpClient("A3Client").AddPolicyHandler(localTimeout.WrapAsync(localRetryPolicy));