+++ StabilityMatrix.Avalonia/Models/OpenArtCustomNode.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace StabilityMatrix.Avalonia.Models;
+
+public class OpenArtCustomNode
+{
+    public required string Title { get; set; }
+    public List<string> Children { get; set; } = [];
+    public bool IsInstalled { get; set; }
+}