+++ StabilityMatrix.Avalonia/ViewModels/CheckpointBrowser/CheckpointBrowserCardViewModel.cs
@@ -19,6 +19,7 @@ using StabilityMatrix.Avalonia.ViewModels.Dialogs;
 using StabilityMatrix.Avalonia.Views.Dialogs;
 using StabilityMatrix.Core.Attributes;
 using StabilityMatrix.Core.Extensions;
+using StabilityMatrix.Core.Helper;
 using StabilityMatrix.Core.Models;
 using StabilityMatrix.Core.Models.Api;
 using StabilityMatrix.Core.Models.Database;
@@ -216,7 +217,7 @@ public partial class CheckpointBrowserCardViewModel : Base.ProgressViewModel
             MaxDialogHeight = 950,
         };
 
-        var prunedDescription = PruneDescription(model);
+        var prunedDescription = Utilities.RemoveHtml(model.Description);
 
         var viewModel = dialogFactory.Get<SelectModelVersionViewModel>();
         viewModel.Dialog = dialog;
@@ -261,23 +262,6 @@ public partial class CheckpointBrowserCardViewModel : Base.ProgressViewModel
 
         await Task.Delay(100);
         await DoImport(model, downloadPath, selectedVersion, selectedFile);
-    }
-
-    private static string PruneDescription(CivitModel model)
-    {
-        var prunedDescription =
-            model
-                .Description?.Replace("<br/>", $"{Environment.NewLine}{Environment.NewLine}")
-                .Replace("<br />", $"{Environment.NewLine}{Environment.NewLine}")
-                .Replace("</p>", $"{Environment.NewLine}{Environment.NewLine}")
-                .Replace("</h1>", $"{Environment.NewLine}{Environment.NewLine}")
-                .Replace("</h2>", $"{Environment.NewLine}{Environment.NewLine}")
-                .Replace("</h3>", $"{Environment.NewLine}{Environment.NewLine}")
-                .Replace("</h4>", $"{Environment.NewLine}{Environment.NewLine}")
-                .Replace("</h5>", $"{Environment.NewLine}{Environment.NewLine}")
-                .Replace("</h6>", $"{Environment.NewLine}{Environment.NewLine}") ?? string.Empty;
-        prunedDescription = HtmlRegex().Replace(prunedDescription, string.Empty);
-        return prunedDescription;
     }
 
     private static async Task<FilePath> SaveCmInfo(